Data memory: two CY7C1041GN-10 (256K x 16, 1 MB); instruction memory IS61C64AL-10
- netlist::Sram16: 256K x 16 single-port SRAM with byte enables (two
  byte-lane cores of the as7c164a model, which now carries 18 address
  lines), TSOP II-44 pin map from the CY7C1041G datasheet.
- as7c164a::Timing::cy7c1041g_10 (datasheet 001-91368, 10 ns column) and
  is61c64al_10 (IS61C256AH-10 column as the stand-in).
- Data memory: MR[19:2] addresses 1 MB; BHE#/BLE# tied low for now.
- Sweep: CY7C1041G-10 clean from 33 ns with the 8 ns gate tap; 136 GALs,
  152 chips plus the gate.  Docs and diagrams updated.

+++ tests/as7c164a.rs
@@ -19,7 +19,7 @@ impl Bench {
     fn new() -> Self {
         let mut chip = As7c164a::new();
         for a in 0..8192u16 {
-            chip.preload(a, (a as u8) ^ 0xA5);
+            chip.preload(a as u32, (a as u8) ^ 0xA5);
         }
         let inp = Inputs::idle();
         chip.set_inputs(0, inp);
